+++ Library/CMakeLists.txt
@@ -58,6 +58,7 @@ set(PLAYRHO_PMR_SRCS
 set(PLAYRHO_D2_HDRS
 	"${PLAYRHO_INCLUDE_PREFIX}d2/AABB.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/Acceleration.hpp"
+	"${PLAYRHO_INCLUDE_PREFIX}d2/BasicAPI.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/Body.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/BodyConf.hpp"
 	"${PLAYRHO_INCLUDE_PREFIX}d2/BodyConstraint.hpp"
